Very dirty refactor
At least, this fixes the problem due to which the tests would hang instead
of terminating: Before, the `single` mode would also try to draw the
pipeline from cache. However, that cache was never inited in that mode
(and shouln'd be in that mode). Now, workflow is a lot cleaner, but the
implementation is still in prototype style.